Various fixes getting Weave working
Some CNI plugins put binaries in /opt/cni/bin, but ours
are in /usr/libexec/cni (default cri-o location). I created
a symlink to the latter and weave started working. Also
had to change the default pod cidr, 192.168.0.0/16, to
something less likely to be in use on a home users network,
10.123.0.0/16.